removed logging dependency
This project will be henceforth a game and no longer simply an engine
for one. Future commits will include implementations of the game instead
of just boilerplate. The game no longer runs due to the usage of an
XML library. The use of this library will be discontinued in future
freleases. Also added tags file for use in vim.